requirements.txt
python-telegram-bot==5.3.0
requests==2.12.5
configparser==3.5.0